+++ src/lib/components/snapps/utmParams.svelte
@@ -1,12 +1,17 @@
 <script lang="ts">
-	import type { SvelteMap } from 'svelte/reactivity';
+	import type { Setting } from '@prisma/client';
 
-	import * as Table from '$lib/components/ui/table/index.js';
+	import { page } from '$app/state';
+	import * as Popover from '$lib/components/ui/popover';
+	import * as Table from '$lib/components/ui/table';
 	import { getTranslations } from '$lib/i18n/index.svelte';
+	import { toast } from 'svelte-sonner';
+	import { SvelteMap } from 'svelte/reactivity';
 
 	import P from '../typography/text/p.svelte';
 	import { Button } from '../ui/button';
 	import { Input } from '../ui/input';
+	import PopoverContent from '../ui/popover/popover-content.svelte';
 
 	const {
 		params = $bindable(),
@@ -25,6 +30,45 @@
 		showHelper?: boolean;
 	} = $props();
 	const i18n = getTranslations();
+
+	const loadGlobalParams = async () => {
+		const f = page.data.fetch as typeof fetch;
+		try {
+			const response =
+				((await (
+					await f(
+						`/api/setting/findFirst?q=${JSON.stringify({
+							where: {
+								field: `GLOBAL_UTM_PARAMS_${page.data.user.id}`,
+								id: `GLOBAL_UTM_PARAMS_${page.data.user.id}`
+							}
+						})}`
+					)
+				).json()) as { data: null | Setting }) || null;
+
+			if ('data' in response && response.data !== null && response.data.value !== '{}') {
+				const stringifiedParams =
+					(response.data.value && response.data.value.trim() !== '' && response.data.value) || '[]';
+				const storedParams = JSON.parse(stringifiedParams || '[]') as string[][];
+
+				for (const [key, value, name] of storedParams) {
+					savedParams.set(`${key}`, { key, name, value });
+				}
+			}
+		} catch (error) {
+			console.error(error);
+			toast.error('errors.generic');
+		}
+	};
+
+	let savedParams: SvelteMap<
+		string,
+		{
+			key: string;
+			name: string;
+			value: string;
+		}
+	> = new SvelteMap();
 </script>
 
 <div class="mb-4 grid content-start gap-0">
@@ -37,10 +81,48 @@
 	<Table.Root>
 		<Table.Header>
 			<Table.Row>
-				{#if showHelper}<Table.Head class="h-10 w-10 p-1"></Table.Head>{/if}
+				{#if showHelper}
+					<Table.Head class="h-10 w-10 p-1"></Table.Head>
+				{/if}
 				<Table.Head class="h-10 p-1 px-2">{i18n.t('snapps.labels.utm.key')}</Table.Head>
 				<Table.Head class="h-10 p-1 px-2">{i18n.t('snapps.labels.utm.value')}</Table.Head>
 				<Table.Head class="h-10 p-1 px-2">{i18n.t('snapps.labels.utm.name')}</Table.Head>
+				{#if showHelper}
+					<Table.Head class="h-10 w-10 p-1">
+						<Popover.Root>
+							<Popover.Trigger>
+								{#snippet child({ props })}
+									<Button {...props} class=" relative h-8 w-8"
+										><i class="ph-duotone ph-list-magnifying-glass text-[24px]"></i></Button
+									>
+								{/snippet}
+							</Popover.Trigger>
+							<PopoverContent class="p-2" align="end">
+								{#await loadGlobalParams()}
+									<div class="flex h-full min-h-[100px] w-full items-center justify-center">
+										<div class="h-5 w-5 animate-spin duration-1000">
+											<i class="ph ph-spinner text-[20px]"></i>
+										</div>
+									</div>
+								{:then}
+									<div class="flex flex-col gap-2 min-h-[100px]">
+										{#each savedParams.entries() as [,param]}
+										<Button class="h-8 p-1 flex items-center gap-2" variant=ghost onclick={()=>{
+											if(params.has(param.key)) params.delete(param.key)
+											else params.set(param.key,param)
+										}}>
+										<i class="ph ph-check transition-all text-[20px] opacity-0" class:opacity-100={params.has(param.key)}></i>
+										<span class="w-full text-start">
+											{param.name}
+										</span>
+									</Button>
+										{/each}
+										</div>
+								{/await}
+							</PopoverContent>
+						</Popover.Root>
+					</Table.Head>
+				{/if}
 			</Table.Row>
 		</Table.Header>
 		<Table.Body>
@@ -80,7 +162,7 @@
 							}}
 						/>
 					</Table.Cell>
-					<Table.Cell class="p-1">
+					<Table.Cell class="p-1" colspan={2}>
 						<Input
 							disabled={param.key === 'utm_' || !showHelper}
 							type="text"
@@ -114,7 +196,7 @@
 					{/if}
 					<Table.Cell class="p-1"></Table.Cell>
 					<Table.Cell class="p-1"></Table.Cell>
-					<Table.Cell class="p-1"></Table.Cell>
+					<Table.Cell class="p-1" colspan={2}></Table.Cell>
 				</Table.Row>
 			</Table.Footer>
 		{/if}