+++ src/lib/server/params/index.ts
@@ -79,3 +79,5 @@ export class ParamsHandler {
 		}
 	}
 }
+
+