+++ src/routes/+page.server.ts
@@ -1,7 +1,11 @@
 import { redirect } from '@sveltejs/kit';
 import { dev } from '$app/environment';
+import { env } from '$env/dynamic/private';
+import { env as pubEnv } from '$env/dynamic/public';
 import { loadTranslations } from '$lib/i18n/server.js';
 import { getSettings } from '$lib/server/config/index.js';
+import { log } from '$lib/server/log/index.js';
+import { getUmami } from '$lib/umami.js';
 import * as shiki from 'shiki'
 export const load = async (event) => {
 	const settings = await getSettings();
@@ -10,12 +14,22 @@ export const load = async (event) => {
 	if (disableHome) redirect(302, '/dashboard');
 	const availableLanguages = settings.get<string>("AVAILABLE_LANGUAGES") as string
 
+	try {
+		const UMAMI_WEBSITE_ID = settings.get<string>('PUBLIC_UMAMI_WEBSITE_ID') ||pubEnv.PUBLIC_UMAMI_WEBSITE_ID
+		const UMAMI_WEBSITE_URL = settings.get<string>('PUBLIC_UMAMI_WEBSITE_URL') ||pubEnv.PUBLIC_UMAMI_WEBSITE_URL
+		const userAgent = event.request.headers.get('user-agent')?.toString()||undefined
+		const umami = getUmami(UMAMI_WEBSITE_URL,UMAMI_WEBSITE_ID,userAgent)
+		await umami.track({title:"/", url:event.url})
+	} catch (error) {
+		if(env.LOG_LEVEL==='debug')log.error(error)
+	}
+
 	return {
 		availableLanguages,
 		disableHome,
 		dockerCompose: await dockerCompose(event.locals.theme),
 		locale: event.locals.lang, 
-				startDocker: await startDocker(event.locals.theme),
+		startDocker: await startDocker(event.locals.theme),
 		translations: await loadTranslations(event.locals.lang||'en')
 	};
 };
@@ -72,4 +86,5 @@ const dockerCompose = async (theme: string) =>
 			lang: 'bash',
 			theme: (theme === 'system' || theme === 'dark' ? 'github-dark' : 'github-light')
 		}
+
 	);