+++ src/routes/[shortcode]/+page.server.ts
@@ -6,11 +6,12 @@ import { singleSchema } from '$lib/components/snapps/schema';
 import { prisma } from '$lib/db/prisma';
 import { log } from '$lib/server/log';
 import { markUsage } from '$lib/server/snapps/markUsage';
+import {  logSecretInvalidOnSnapp, logSnappNotFound } from '$lib/umami';
 import bcrypt from 'bcryptjs';
 import { superValidate } from 'sveltekit-superforms';
 import { zod } from 'sveltekit-superforms/adapters';
 
-import type { PageServerLoad } from './$types';
+import type { PageServerLoad, } from './$types';
 
 export const load: PageServerLoad = async (event) => {
 	const {
@@ -26,7 +27,10 @@ export const load: PageServerLoad = async (event) => {
 
 
 
-	if (!snapp) throw error(404, { message: 'errors.snapps.not-found' });
+	if (!snapp) {
+		logSnappNotFound(event)
+		throw error(404, { message: 'errors.snapps.not-found' });
+	}
 	const url = new URL(snapp.originalUrl);
 
 	const utmParamsString = JSON.parse(snapp.utmParams || '[]') as string[];
@@ -40,7 +44,7 @@ export const load: PageServerLoad = async (event) => {
 	}
 	if (snapp.secret === null && snapp.userId !== user?.id) {
 		const [available, err] = await markUsage(event, snapp);
-		
+
 		if (available) redirect(302, encodeURI(decodeURI(url.toString())));
 		else {
 			return {
@@ -56,7 +60,7 @@ export const load: PageServerLoad = async (event) => {
 
 	return {
 		form: await superValidate(zod(singleSchema)),
-		hasPassword: snapp.secret!==null,
+		hasPassword: snapp.secret !== null,
 		isDisabled: snapp.disabled
 	};
 };
@@ -80,34 +84,41 @@ export const actions = {
 			log.error(error);
 			return { form: secretForm, message: "errors.generic" }
 		}
-			if (!snapp) return fail(400, { form: secretForm, message: 'errors.snapps.not-found', });
-			const isPasswordCorrect = await bcrypt.compare(secretForm.data.secret, snapp.secret!)
+		if (!snapp) {
+			logSnappNotFound(event)
+			return fail(400, { form: secretForm, message: 'errors.snapps.not-found', });
+		}
+		const isPasswordCorrect = await bcrypt.compare(secretForm.data.secret, snapp.secret!)
 
-			if (!isPasswordCorrect) return fail(400, { form: secretForm, message: "errors.auth.wrong-credentials" })
-			const [available, err] = await markUsage(event, snapp);
-
-			const url = new URL(snapp.originalUrl);
-
-			const utmParamsString = JSON.parse(snapp.utmParams || '[]') as string[];
-			const utmParams = utmParamsString.map((p) => {
-				const [key, value, name] = JSON.parse(p) as string[];
-				return { key, name, value };
-			});
-
-			for (const params of utmParams) {
-				url.searchParams.set(params.key, params.value);
-			}
-
-			if (available) redirect(302, url);
-			else 
-				return {
-			err,
-			form:secretForm,
-					hasPassword: false,
-					isDisabled: true,
-					message:'errors.snapps.disabled',
-				};
-			
+		if (!isPasswordCorrect) {
+			logSecretInvalidOnSnapp(event)
+			return fail(400, { form: secretForm, message: "errors.auth.wrong-credentials" })
+		}
 		
+		const [available, err] = await markUsage(event, snapp);
+
+		const url = new URL(snapp.originalUrl);
+
+		const utmParamsString = JSON.parse(snapp.utmParams || '[]') as string[];
+		const utmParams = utmParamsString.map((p) => {
+			const [key, value, name] = JSON.parse(p) as string[];
+			return { key, name, value };
+		});
+
+		for (const params of utmParams) {
+			url.searchParams.set(params.key, params.value);
+		}
+
+		if (available) redirect(302, url);
+		else
+			return {
+				err,
+				form: secretForm,
+				hasPassword: false,
+				isDisabled: true,
+				message: 'errors.snapps.disabled',
+			};
+
+
 	}
 };