+++ src/routes/auth/sign-in/+page.server.ts
@@ -6,16 +6,18 @@ import { createSession } from '$lib/server/auth/index.js';
 import { setSessionTokenCookie } from '$lib/server/auth/index.js';
 import { OIDCConfigs } from '$lib/server/auth/oidc/config';
 import { watchLists } from '$lib/server/watchlists';
+import { logInvalidLoginAttempt } from '$lib/umami';
 import bcrypt from 'bcryptjs';
 import { superValidate } from 'sveltekit-superforms';
 import { zod } from 'sveltekit-superforms/adapters';
+
 export const load = async ({ locals: { user } }) => {
 	if (user) redirect(302, '/dashboard');
 	const providers = OIDCConfigs.map((o) => ({ identity: o.identity }));
-	const DISABLED_EMAIL_AND_PASSWORD  = process.env.DISABLED_EMAIL_AND_PASSWORD?.toLowerCase() === "true" || false
-	
+	const DISABLED_EMAIL_AND_PASSWORD = process.env.DISABLED_EMAIL_AND_PASSWORD?.toLowerCase() === "true" || false
+
 	return {
-		emailDisabled:DISABLED_EMAIL_AND_PASSWORD,
+		emailDisabled: DISABLED_EMAIL_AND_PASSWORD,
 		form: await superValidate(zod(signInSchema)),
 		providers
 	};
@@ -37,7 +39,7 @@ export const actions = {
 			return fail(400, { form, message: 'errors.auth.user-not-found' });
 		}
 
-		const disabled = process.env.DISABLED_EMAIL_AND_PASSWORD?.toLowerCase() === 'true'||false
+		const disabled = process.env.DISABLED_EMAIL_AND_PASSWORD?.toLowerCase() === 'true' || false
 		if (disabled) {
 			return fail(401, { form, message: 'errors.auth.disabled-auth-email-and-password' });
 		}
@@ -45,15 +47,17 @@ export const actions = {
 		const validEmail = await watchLists.checkEmail(auth.email);
 		const validUsername = await watchLists.checkUsername(form.data.username);
 
-		if (!validEmail || !validUsername) {
+		if (!validEmail || !validUsername)
 			return fail(400, { form, message: 'errors.blacklisted.user' });
-		}
 
 		const validPassword = bcrypt.compareSync(form.data.password, auth.password);
 
 		if (!validPassword) {
+			logInvalidLoginAttempt(event)
 			return fail(400, { form, message: 'errors.auth.wrong-credentials' });
 		}
+
+
 		const sessionId = generateSessionToken();
 		const session = await createSession(sessionId, auth.id);
 
@@ -69,3 +73,4 @@ export const actions = {
 		};
 	}
 };
+