+++ src/routes/db-error/+page.server.ts
@@ -3,12 +3,15 @@ import { getSettings } from '$lib/server/config';
 import DbErrorEmail from '$lib/server/emails/dbErrorEmail.svelte';
 import { log } from '$lib/server/log';
 import { sendEmail } from '$lib/server/smtp';
+import { logDatabaseNotAvailable } from '$lib/umami';
+
+
 
 const hasSentEmail = {
 	sent: null as Date | null
 };
 
-export const load = async () => {
+export const load = async (event) => {
 	const settings = await getSettings();
 	const DATABASE_OFFLINE = settings.get<boolean>('DB_OFFLINE');
 	if (!DATABASE_OFFLINE) redirect(302, '/dashboard');
@@ -20,7 +23,9 @@ export const load = async () => {
 			await sendEmail(DbErrorEmail, {}, process.env.ADMIN_EMAIL || '', 'SNAPP DB OFFLINE');
 			hasSentEmail.sent = new Date();
 		}
+		logDatabaseNotAvailable(event)
 	} catch (error) {
 		if (process.env.LOG_LEVEL === 'debug') log.error(error);
 	}
 };
+