+++ src/routes/groups/[groupId]/[shortcode]/+page.server.ts
@@ -6,6 +6,7 @@ import { singleSchema } from '$lib/components/snapps/schema';
 import { prisma } from '$lib/db/prisma';
 import { log } from '$lib/server/log';
 import { markUsage } from '$lib/server/snapps/markUsage';
+import { logSecretInvalidOnSnapp, logSnappNotFound } from '$lib/umami';
 import bcrypt from 'bcryptjs';
 import { superValidate } from 'sveltekit-superforms';
 import { zod } from 'sveltekit-superforms/adapters';
@@ -25,7 +26,10 @@ export const load: PageServerLoad = async (event) => {
 	})
 
 
-	if (!snapp) throw error(404, { message: 'errors.snapps.not-found' });
+	if (!snapp) {
+		logSnappNotFound(event)
+		throw error(404, { message: 'errors.snapps.not-found' });
+	}
 	const url = new URL(snapp.originalUrl);
 
 	const utmParamsString = JSON.parse(snapp.utmParams || '[]') as string[];
@@ -60,51 +64,58 @@ export const load: PageServerLoad = async (event) => {
 
 export const actions = {
 	trySecret: async (event) => {
-			const {
-				params: { shortcode }
-			} = event;
-			const secretForm = await superValidate(event, zod(singleSchema));
-	
-			if (!secretForm.valid) {
-				return fail(400, {
-					form: secretForm
-				});
-			}
-			let snapp: null | Snapp = null;
-			try {
-				snapp = await prisma.snapp.findFirst({ where: { shortcode } });
-			} catch (error) {
-				log.error(error);
-				return { form: secretForm, message: "errors.generic" }
-			}
-				if (!snapp) return fail(400, { form: secretForm, message: 'errors.snapps.not-found', });
-	
-				const isPasswordCorrect = await bcrypt.compare(secretForm.data.secret, snapp.secret!)
-	
-				if (!isPasswordCorrect) return fail(400, { form: secretForm, message: "errors.auth.wrong-credentials" })
-				const [available, err] = await markUsage(event, snapp);
-	
-				const url = new URL(snapp.originalUrl);
-	
-				const utmParamsString = JSON.parse(snapp.utmParams || '[]') as string[];
-				const utmParams = utmParamsString.map((p) => {
-					const [key, value, name] = JSON.parse(p) as string[];
-					return { key, name, value };
-				});
-	
-				for (const params of utmParams) {
-					url.searchParams.set(params.key, params.value);
-				}
-	
-				if (available) redirect(302, url);
-				else {
-					return {
-						err,
-						hasPassword: false,
-						isDisabled: true,
-						message:undefined
-					};
-				}
-			
+		const {
+			params: { shortcode }
+		} = event;
+		const secretForm = await superValidate(event, zod(singleSchema));
+
+		if (!secretForm.valid) {
+			return fail(400, {
+				form: secretForm
+			});
 		}
+		let snapp: null | Snapp = null;
+		try {
+			snapp = await prisma.snapp.findFirst({ where: { shortcode } });
+		} catch (error) {
+			log.error(error);
+			return { form: secretForm, message: "errors.generic" }
+		}
+		if (!snapp) {
+			logSnappNotFound(event)
+			return fail(400, { form: secretForm, message: 'errors.snapps.not-found', });}
+
+		const isPasswordCorrect = await bcrypt.compare(secretForm.data.secret, snapp.secret!)
+
+		if (!isPasswordCorrect) {
+			logSecretInvalidOnSnapp(event)
+			return fail(400, { form: secretForm, message: "errors.auth.wrong-credentials" })
+		}
+		const [available, err] = await markUsage(event, snapp);
+
+		const url = new URL(snapp.originalUrl);
+
+		const utmParamsString = JSON.parse(snapp.utmParams || '[]') as string[];
+		const utmParams = utmParamsString.map((p) => {
+			const [key, value, name] = JSON.parse(p) as string[];
+			return { key, name, value };
+		});
+
+		for (const params of utmParams) {
+			url.searchParams.set(params.key, params.value);
+		}
+
+		if (available) redirect(302, url);
+		else {
+			return {
+				err,
+				hasPassword: false,
+				isDisabled: true,
+				message: undefined
+			};
+		}
+
+	}
 };
+
+